requirements.txt
aiogram==3.4.1
sqlmodel==0.0.16
python-dotenv==1.0.1
openai>=1.0,<2.0
httpx<0.25
backports.zoneinfo; python_version < "3.9"
apscheduler>=3.10